@@ -9,33 +9,26 @@
     "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. src/"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "3.0.1",
-    "@sveltejs/kit": "2.5.9",
-    "@sveltejs/vite-plugin-svelte": "^3.1.0",
-    "@svgdotjs/svg.js": "3.2.0",
+    "@sveltejs/adapter-static": "3.0.6",
+    "@sveltejs/kit": "2.7.3",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@svgdotjs/svg.js": "3.2.4",
     "clipboard": "2.0.11",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-svelte3": "4.0.0",
-    "highlight.js": "11.9.0",
+    "eslint-plugin-svelte": "^2.46.0",
+    "highlight.js": "11.10.0",
     "normalize.css": "^8.0.1",
-    "postcss": "8.4.38",
+    "postcss": "8.4.47",
     "postcss-import": "16.1.0",
-    "postcss-load-config": "5.1.0",
-    "postcss-preset-env": "9.5.13",
+    "postcss-load-config": "6.0.1",
+    "postcss-preset-env": "10.0.8",
     "postcss-reporter": "^7.1.0",
-    "prettier": "3.2.5",
-    "prettier-plugin-svelte": "3.2.3",
-    "svelte": "4.2.17",
-    "svelte-preprocess": "5.1.4",
-    "vite": "5.2.11"
+    "prettier": "3.3.3",
+    "prettier-plugin-svelte": "3.2.7",
+    "svelte": "5.1.3",
+    "svelte-preprocess": "6.0.3",
+    "vite": "5.4.10"
   },
-  "type": "module",
-  "pnpm": {
-    "peerDependencyRules": {
-      "ignoreMissing": [
-        "eslint-plugin-svelte3"
-      ]
-    }
-  }
+  "type": "module"
 }
