@@ -1,8 +1,8 @@
 module.exports = {
   root: true,
-  extends: ['eslint:recommended', 'prettier'],
-  plugins: ['svelte3'],
-  overrides: [{files: ['*.svelte'], processor: 'svelte3/svelte3'}],
+  extends: ['eslint:recommended', 'prettier', 'plugin:svelte/recommended'],
+  plugins: [],
+  overrides: [{ files: ['*.svelte'], parser: 'svelte-eslint-parser'}],
   parserOptions: {
     sourceType: 'module',
     ecmaVersion: 2020
